Crate argh_shared

Shared functionality between argh_derive and the argh runtime.

This library is intended only for internal use by these two crates.

Structs

CommandInfo

Information about a particular command used for output.

CommandInfoWithArgs

Information about the command line arguments for a given command.

ErrorCodeInfo

Information about a documented error code.

FlagInfo

Information about a flag or option.

PositionalInfo

Information about positional arguments

SubCommandInfo

Information about a subcommand. Dynamic subcommands do not implement get_args_info(), so the command field only contains the name and description.

Enums

FlagInfoKind

The kind of flags.

Optionality

The optionality defines the requirements related to the presence of the argument on the command line.

Constants

INDENT

Functions

write_description

Write command names and descriptions to an output string.